Split up JarOps and JarManifest
